liblights: Support prioritized multiplexing of LEDs.
Support prioritized multiplexing of battery, notifications, and attention
LEDs.  That is, show the LED with highest priority (e.g., attention), and
when that LED is turned off, fall back to the LED with next highest
priority (e.g., battery or notifications).

Also removes use of get_calibrated_color when setting LEDs, and other minor
cleanup.